backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
babel==2.17.0
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.3
click==8.2.1
constantly==23.10.4
courlan==1.3.2
coverage==7.10.4
cryptography==45.0.6
cytoolz==1.0.1
daphne==4.1.2
dateparser==1.2.2
distlib==0.4.0
Django==5.0.6
django-cors-headers==4.4.0
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
dotenv==0.9.9
eth-hash==0.7.1
eth-typing==5.2.1
eth-utils==5.3.0
factory_boy==3.3.3
Faker==37.5.3
filelock==3.19.1
greenlet==3.2.4
h11==0.16.0
htmldate==1.9.3
hyperlink==21.0.0
identify==2.6.13
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
isort==6.0.1
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
msgpack==1.1.1
mypy_extensions==1.1.0
nodeenv==1.9.1
noise==1.2.2
numpy==2.3.2
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
playwright==1.54.0
pluggy==1.6.0
pre_commit==4.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyee==13.0.0
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
PySocks==1.7.1
pytest==8.4.1
pytest-asyncio==1.2.0
pytest-cov==6.2.1
pytest-django==4.11.1
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.4
requests-file==2.1.0
ruff==0.12.9
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
stem==1.8.2
tld==0.13.1
tldextract==5.3.0
toolz==1.0.0
trafilatura==2.0.0
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.34.0
webencodings==0.5.1
websockets==15.0.1
zope.interface==7.2
